requirements.txt
fastapi[standard]==0.111.0
uvicorn==0.30.1
streamlit==1.35.0
langchain==0.2.1
langchain-google-genai==0.0.12
google-generativeai==0.6.2
google-api-python-client==2.130.0
python-dotenv==1.0.1
